adding new file for common utilities
