+++ x-pack/plugins/security_solution/public/entity_analytics/components/risk_summary_flyout/risk_summary.tsx
@@ -6,7 +6,7 @@
  */
 
 import React, { useMemo } from 'react';
-import type { EuiBasicTableColumn } from '@elastic/eui';
+
 import {
   useEuiTheme,
   EuiAccordion,
@@ -20,12 +20,9 @@ import {
 import { css } from '@emotion/react';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { euiThemeVars } from '@kbn/ui-theme';
-import { i18n } from '@kbn/i18n';
+
 import { EntityDetailsLeftPanelTab } from '../../../flyout/entity_details/shared/components/left_panel/left_panel_header';
-import type {
-  HostRiskScore,
-  UserRiskScore,
-} from '../../../../common/search_strategy/security_solution/risk_score';
+
 import { InspectButton, InspectButtonContainer } from '../../../common/components/inspect';
 import { ONE_WEEK_IN_HOURS } from '../../../timelines/components/side_panel/new_user_detail/constants';
 import { FormattedRelativePreferenceDate } from '../../../common/components/formatted_date';
@@ -34,37 +31,20 @@ import { VisualizationEmbeddable } from '../../../common/components/visualizatio
 import { ExpandablePanel } from '../../../flyout/shared/components/expandable_panel';
 import type { RiskScoreState } from '../../api/hooks/use_risk_score';
 import { getRiskScoreSummaryAttributes } from '../../lens_attributes/risk_score_summary';
+import {
+  buildColumns,
+  getEntityData,
+  getItems,
+  isUserRiskData,
+  LAST_30_DAYS,
+  LENS_VISUALIZATION_HEIGHT,
+} from './common';
 
 export interface RiskSummaryProps<T extends RiskScoreEntity> {
   riskScoreData: RiskScoreState<T>;
   queryId: string;
   openDetailsPanel: (tab: EntityDetailsLeftPanelTab) => void;
 }
-
-interface TableItem {
-  category: string;
-  count: number;
-}
-const LENS_VISUALIZATION_HEIGHT = 126; //  Static height in pixels specified by design
-const LAST_30_DAYS = { from: 'now-30d', to: 'now' };
-
-function isUserRiskData(
-  riskData: UserRiskScore | HostRiskScore | undefined
-): riskData is UserRiskScore {
-  return !!riskData && (riskData as UserRiskScore).user !== undefined;
-}
-
-const getEntityData = (riskData: UserRiskScore | HostRiskScore | undefined) => {
-  if (!riskData) {
-    return;
-  }
-
-  if (isUserRiskData(riskData)) {
-    return riskData.user;
-  }
-
-  return riskData.host;
-};
 
 const RiskSummaryComponent = <T extends RiskScoreEntity>({
   riskScoreData,
@@ -88,50 +68,10 @@ const RiskSummaryComponent = <T extends RiskScoreEntity>({
     });
   }, [entityData?.name, entityData?.risk?.calculated_level, riskData]);
 
-  const columns: Array<EuiBasicTableColumn<TableItem>> = useMemo(
-    () => [
-      {
-        field: 'category',
-        name: (
-          <FormattedMessage
-            id="xpack.securitySolution.flyout.entityDetails.categoryColumnLabel"
-            defaultMessage="Category"
-          />
-        ),
-        truncateText: false,
-        mobileOptions: { show: true },
-        sortable: true,
-      },
-      {
-        field: 'count',
-        name: (
-          <FormattedMessage
-            id="xpack.securitySolution.flyout.entityDetails.inputsColumnLabel"
-            defaultMessage="Inputs"
-          />
-        ),
-        truncateText: false,
-        mobileOptions: { show: true },
-        sortable: true,
-        dataType: 'number',
-      },
-    ],
-    []
-  );
-
   const xsFontSize = useEuiFontSize('xxs').fontSize;
 
-  const items: TableItem[] = useMemo(
-    () => [
-      {
-        category: i18n.translate('xpack.securitySolution.flyout.entityDetails.alertsGroupLabel', {
-          defaultMessage: 'Alerts',
-        }),
-        count: entityData?.risk.inputs?.length ?? 0,
-      },
-    ],
-    [entityData?.risk.inputs?.length]
-  );
+  const columns = useMemo(buildColumns, []);
+  const rows = useMemo(() => getItems(entityData), [entityData]);
 
   return (
     <EuiAccordion
@@ -259,7 +199,7 @@ const RiskSummaryComponent = <T extends RiskScoreEntity>({
                   data-test-subj="risk-summary-table"
                   responsive={false}
                   columns={columns}
-                  items={items}
+                  items={rows}
                   compressed
                 />
               </div>
